Change directory structure to benefit from explicit file names
