+++ src/index.js
@@ -1,8 +1,8 @@
 /* @flow */
 import injectGlobal from './assets/styles/global';
 
-import Button from './components/Button';
-import ButtonGroup from './components/ButtonGroup';
-import Label from './components/Label';
+import { Button, ButtonGroup } from './components/Button';
+import { Label } from './components/Label';
+import { Segment } from './components/Segment';
 
-export { injectGlobal, Button, ButtonGroup, Label };
+export { injectGlobal, Button, ButtonGroup, Label, Segment };
